requirements.txt
Flask==2.0.3
tensorflow==2.8.0
numpy==1.21.5
Pillow==9.0.1
matplotlib==3.5.1
tqdm==4.62.3
nltk==3.7
keras==2.8.0
